Improves card view and adds overlay functionality
Refactors the CardsView to improve its layout and add a button for displaying an overlay.

The overlay functionality allows the user to add new cards through the search overlay view. The CardsView now uses a binding to control the presentation of the search overlay.
The Profile view is updated to use the general showOverlay.
SearchOverlayView no longer needs a binding.